+++ include/Geometry/ParameterDesign/Interpolation.h
@@ -13,7 +13,7 @@ enum class Result
 class Interpolation
 {
 public:
-	Interpolation(const std::vector<Point2f>& points) :Points(points) {  }
+	Interpolation(const std::vector<Point2d>& points) :Points(points) {  }
 
 	virtual Result error_bound(Float& error_b) = 0;
 
@@ -22,7 +22,7 @@ public:
 
 	virtual  Float operator() (Float in_val) = 0;
 protected:
-	std::vector<Point2f> Points;
+	std::vector<Point2d> Points;
 };
 
 class RadialInterpolation : public Interpolation
@@ -38,8 +38,8 @@ public:
 		{
 			if (Points.size() > 1)
 			{
-				auto iter_max = std::max_element(Points.begin(), Points.end(), [](const Point2f& a, const Point2f& b) {return a.x() < b.x(); });
-				auto iter_min = std::min_element(Points.begin(), Points.end(), [](const Point2f& a, const Point2f& b) {return a.x() < b.x(); });
+				auto iter_max = std::max_element(Points.begin(), Points.end(), [](const Point2d& a, const Point2d& b) {return a.x() < b.x(); });
+				auto iter_min = std::min_element(Points.begin(), Points.end(), [](const Point2d& a, const Point2d& b) {return a.x() < b.x(); });
 				d = (iter_max->x() - iter_min->x()) / Points.size();
 				d = d * d;
 			}
@@ -154,7 +154,7 @@ public:
 	void evaluate() override;
 	Float operator()(Float in_val) override;
 
-	void addPoint(const Point2f& point)
+	void addPoint(const Point2d& point)
 	{
 		Points.push_back(point);
 	}
@@ -206,14 +206,14 @@ class HermitePolynomial :public Interpolation
 {
 public:
 
-	HermitePolynomial(const std::vector<Point2f>& points, const std::vector<Point2f>& derivatives)
+	HermitePolynomial(const std::vector<Point2d>& points, const std::vector<Point2d>& derivatives)
 		:Interpolation(points), derivatives(derivatives) {	}
 	Result error_bound(Float& error_b) override;
 	void evaluate() override;
 	Float operator()(Float in_val) override;
 
 private:
-	std::vector<Point2f> derivatives;
+	std::vector<Point2d> derivatives;
 };
 
 class SplineInterpolation :public Interpolation
@@ -229,7 +229,7 @@ public:
 class BezierSplineInterpolation :public Interpolation
 {
 public:
-	BezierSplineInterpolation(const std::vector<Point2f>& points)
+	BezierSplineInterpolation(const std::vector<Point2d>& points)
 		: Interpolation(points)
 	{
 	}
@@ -255,7 +255,7 @@ public:
 		int mat_size = point_count * 3 - 2;
 		matrix = Eigen::SparseMatrix<Float>(mat_size, mat_size);
 
-		std::sort(Points.begin(), Points.end(), [](const Point2f& a, const Point2f& b) {return a.x() < b.x(); });
+		std::sort(Points.begin(), Points.end(), [](const Point2d& a, const Point2d& b) {return a.x() < b.x(); });
 
 		Interval interval(Points[0].x(), Points.back().x());
 		std::vector<Eigen::Triplet<Float>> triplets;
@@ -344,13 +344,13 @@ public:
 				{
 					return rst[0];
 				}
-				std::vector<Point2f> bezier_points(4);
+				std::vector<Point2d> bezier_points(4);
 
 				auto diff = Points[i + 1].x() - Points[i].x();
 
 				for (int j = 0; j < 4; ++j)
 				{
-					bezier_points[j] = Point2f(Points[i].x() + j * (diff) / 3.0, rst[3 * i + j]);
+					bezier_points[j] = Point2d(Points[i].x() + j * (diff) / 3.0, rst[3 * i + j]);
 				}
 
 				BezierApproximation bezier(bezier_points);
@@ -366,13 +366,13 @@ public:
 
 	Vector rst;
 
-	std::vector<std::vector<Point2f>> parts;
+	std::vector<std::vector<Point2d>> parts;
 };
 
 class BSplineInterpolation :public Interpolation
 {
 public:
-	explicit BSplineInterpolation(const std::vector<Point2f>& matrices)
+	explicit BSplineInterpolation(const std::vector<Point2d>& matrices)
 		: Interpolation(matrices)
 	{
 	}
@@ -388,7 +388,7 @@ public:
 		{
 			return;
 		}
-		std::sort(Points.begin(), Points.end(), [](const Point2f& a, const Point2f& b) {return a.x() < b.x(); });
+		std::sort(Points.begin(), Points.end(), [](const Point2d& a, const Point2d& b) {return a.x() < b.x(); });
 
 		interval = Interval(Points.front().x(), Points.back().x());
 		interval.SetPartitionCount(Points.size());
@@ -444,5 +444,5 @@ public:
 
 	Interval interval;
 
-	std::vector<Point2f> DeBoorPoints;
+	std::vector<Point2d> DeBoorPoints;
 };
