+++ src/test/FEM/FEM1D/test.cpp
@@ -318,7 +318,7 @@ protected:
 	void DragPoint();
 	void draw(bool* p_open) override;
 
-	std::vector<Point2f> points;
+	std::vector<Point2d> points;
 	bool updated = true;
 	int segemnt = 1;
