+++ src/test/Interpolation/test.cpp
@@ -14,7 +14,7 @@ protected:
 	void DragPoint();
 	void draw(bool* p_open) override;
 
-	std::vector<Point2f> points;
+	std::vector<Point2d> points;
 	bool updated = true;
 
 	void evaluate_lagrangian()
